+++ tempest_fastapi_sdk/__init__.py
@@ -250,7 +250,7 @@ from tempest_fastapi_sdk.websockets import (
     make_websocket_router,
 )
 
-__version__: str = "0.48.0"
+__version__: str = "0.49.0"
 
 __all__: list[str] = [
     "BASE_COLUMN_ORDER",